Add config flag for remove file before write

+++ codegen/templates/templates.go
@@ -58,6 +58,11 @@ type Options struct {
 
 	// Packages cache, you can find me on config.Config
 	Packages *code.Packages
+
+	// RemoveFileBeforeWrite will remove the file before writing to it.
+	// This is useful where filename case might change, so the file name will actually change
+	// on case-insensitive file systems.
+	RemoveFileBeforeWrite bool
 }
 
 var (
@@ -148,7 +153,7 @@ func Render(cfg Options) error {
 	}
 	CurrentImports = nil
 
-	err = write(cfg.Filename, result.Bytes(), cfg.Packages)
+	err = write(cfg.Filename, result.Bytes(), cfg.Packages, cfg.RemoveFileBeforeWrite)
 	if err != nil {
 		return err
 	}
@@ -664,7 +669,7 @@ func render(filename string, tpldata any) (*bytes.Buffer, error) {
 	return buf, t.Execute(buf, tpldata)
 }
 
-func write(filename string, b []byte, packages *code.Packages) error {
+func write(filename string, b []byte, packages *code.Packages, removeFile bool) error {
 	err := os.MkdirAll(filepath.Dir(filename), 0o755)
 	if err != nil {
 		return fmt.Errorf("failed to create directory: %w", err)
@@ -676,9 +681,11 @@ func write(filename string, b []byte, packages *code.Packages) error {
 		formatted = b
 	}
 
-	err = os.Remove(filename)
-	if err != nil && !errors.Is(err, fs.ErrNotExist) {
-		return err
+	if removeFile {
+		err = os.Remove(filename)
+		if err != nil && !errors.Is(err, fs.ErrNotExist) {
+			return err
+		}
 	}
 
 	err = os.WriteFile(filename, formatted, 0o644)
